docs/tutorials/generative-ai/sql-agent-app/requirements.txt
streamlit
ibm-watsonx-ai 
langchain-ibm
langgraph
langchain_experimental
tavily-python
langchain-community
python-dotenv
image